fix(nodata): unify NoData semantics across operators, dialogs and interop (#444, #445)
- Float-space NoData compares (flag + static_cast<float>) replace un-cast
  double compares with absolute tolerance or exact ==: spectral_index
  operator, change_streaming (3 sites), QUAC scene stats, radiometric
  streaming tiles, PCA/MNF isPixelValid, segment features (3 sites),
  spectral unmixing, spectral_roi stats (#444)
- opencv_utils: honor the hasNoData flag on read (undeclared NoData no
  longer masks valid 0 pixels) and on write (output declared NoData only
  from a declared source value, never fabricated 0; NaN pixels stay NaN
  when undeclared) (#444/#445)
- image_fusion: resolve working NoData from pan then first MS band with
  a declaration; NaN when none — MS NoData no longer ignored (#445)
- spectral_resample: per-band NoData propagation — interpolating across
  a sentinel now yields NaN instead of a bogus mid-value (#445)
- terrain_dialog: drop hardcoded -9999 param so the operator resolves
  the DEM's declared NoData (#445)
- contrast_stretch_dialog + image_enhancement_panel: resolve per-band
  NoData and pass it to every stretch; linear min/max computed over
  valid pixels only; IHS preview masks declared NoData (#445)
- rs_obia_segmentation: NaN sentinel when no NoData declared instead of
  fabricated -9999 (#449 item 3)

Fixes #444
Fixes #445

+++ src/analysis/segmentation/rs_segment_features.cpp
@@ -79,7 +79,7 @@ RsSegmentFeatures::extract( const QString &rasterPath,
         {
             if ( !validMask[i] ) continue;
             const float v = bandBuf[i];
-            if ( std::isnan( v ) || ( hasNoData[b] && static_cast<double>( v ) == bandNoData[b] ) )
+            if ( std::isnan( v ) || ( hasNoData[b] && v == static_cast<float>( bandNoData[b] ) ) )
                 validMask[i] = 0;
         }
     }
@@ -297,7 +297,7 @@ RsSegmentFeatures::extract( const QString &rasterPath,
                     const float val1 = bandBuf[idx];
                     // validMask guarantees not NaN/NoData, but keep guard for safety
                     if ( std::isnan(val1) ) continue;
-                    if ( hasNoData[b] && static_cast<double>(val1) == bandNoData[b] ) continue;
+                    if ( hasNoData[b] && val1 == static_cast<float>( bandNoData[b] ) ) continue;
                     const size_t level1 = static_cast<size_t>( std::clamp( static_cast<int>( ( val1 - minV ) / rangeV * ( nLevels - 1 ) ), 0, nLevels - 1 ) );
                     const int drs[] = { 0, 1 };
                     const int dcs[] = { 1, 0 };
@@ -311,7 +311,7 @@ RsSegmentFeatures::extract( const QString &rasterPath,
                         if ( !validMask[nIdx] ) continue;
                         const float val2 = bandBuf[nIdx];
                         if ( std::isnan(val2) ) continue;
-                        if ( hasNoData[b] && static_cast<double>(val2) == bandNoData[b] ) continue;
+                        if ( hasNoData[b] && val2 == static_cast<float>( bandNoData[b] ) ) continue;
                         const size_t level2 = static_cast<size_t>( std::clamp( static_cast<int>( ( val2 - minV ) / rangeV * ( nLevels - 1 ) ), 0, nLevels - 1 ) );
                         glcm[level1][level2] += 1.0;
                         glcm[level2][level1] += 1.0;
